# backend/docs/swagger.yaml
basePath: /
definitions:
  dto.IngredientParamRequest:
    properties:
      name:
        type: string
      stock:
        type: number
      unit_id:
        type: integer
    type: object
  dto.MenuCreateRequest:
    properties:
      description:
        type: string
      image:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/dto.MenuIngredientRequest'
        type: array
      name:
        type: string
      price:
        type: number
    required:
    - image
    - ingredients
    - name
    - price
    type: object
  dto.MenuIngredientRequest:
    properties:
      ingredient_id:
        type: integer
      quantity:
        type: number
      unit_id:
        type: integer
    required:
    - ingredient_id
    - quantity
    - unit_id
    type: object
  dto.MenuUpdateRequest:
    properties:
      description:
        type: string
      image:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/dto.MenuIngredientRequest'
        type: array
      name:
        type: string
      price:
        type: number
    required:
    - image
    - ingredients
    - name
    - price
    type: object
  dto.UnitParamRequest:
    properties:
      name:
        type: string
      symbol:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for Awis Palace Ingredient Management built with
    Gin and GORM.
  title: Awis Palace Ingredient Management API
  version: "1.0"
paths:
  /ingredients:
    get:
      description: Get Ingredients
      responses: {}
      summary: Get Ingredients
      tags:
      - Ingredients
    post:
      description: Post Ingredients
      parameters:
      - description: Create ingredient
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/dto.IngredientParamRequest'
      responses: {}
      summary: Post Ingredients
      tags:
      - Ingredients
  /ingredients/{id}:
    delete:
      description: Delete a ingredient by ID
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: Delete Ingredient
      tags:
      - Ingredients
    put:
      description: Update an existing ingredient by ID
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated ingredient data
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/dto.IngredientParamRequest'
      responses: {}
      summary: Update Ingredient
      tags:
      - Ingredients
  /menus:
    get:
      description: Get Menus
      responses: {}
      summary: Get Menus
      tags:
      - Menus
      - Menus
    post:
      description: Create Menu
      parameters:
      - description: Create menu
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/dto.MenuCreateRequest'
      responses: {}
      summary: Create Menu
      tags:
      - Menus
      - Menus
  /menus/{id}:
    delete:
      description: Delete Menu by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Delete Menu
      tags:
      - Menus
    put:
      description: Update Menu by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update menu
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/dto.MenuUpdateRequest'
      responses: {}
      summary: Update Menu
      tags:
      - Menus
  /units:
    get:
      description: Get Units
      responses: {}
      summary: Get Units
      tags:
      - Units
    post:
      description: Post Units
      parameters:
      - description: Unit data
        in: body
        name: unit
        required: true
        schema:
          $ref: '#/definitions/dto.UnitParamRequest'
      responses: {}
      summary: Post Units
      tags:
      - Units
  /units/{id}:
    delete:
      description: Delete a unit by ID
      parameters:
      - description: Unit ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: Delete Unit
      tags:
      - Units
    put:
      description: Update an existing unit by ID
      parameters:
      - description: Unit ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated unit data
        in: body
        name: unit
        required: true
        schema:
          $ref: '#/definitions/dto.UnitParamRequest'
      responses: {}
      summary: Update Unit
      tags:
      - Units
  /users:
    get:
      description: Get Users
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Users
      tags:
      - Users
swagger: "2.0"